Refactor ParMapN to enable arity up to 9
